Refactor storage manager
